+++ app/models/user.rb
@@ -1,12 +1,11 @@
 # frozen_string_literal: true
 
-class User < ApplicationRecord
+class User < Contributor
   self.abstract_class = true
 
   attr_reader :current_study_group_id
 
   belongs_to :consumer
-  has_many :anomaly_notifications, as: :contributor, dependent: :destroy
   has_many :authentication_token, dependent: :destroy
   has_many :comments, as: :user
   has_many :study_group_memberships, as: :user
@@ -15,23 +14,18 @@ class User < ApplicationRecord
   has_many :programming_groups, through: :programming_group_memberships, as: :user
   has_many :exercises, as: :user
   has_many :file_types, as: :user
-  has_many :submissions, as: :contributor
   has_many :participations, through: :submissions, source: :exercise, as: :user
   has_many :user_proxy_exercise_exercises, as: :user
-  has_many :user_exercise_interventions, as: :contributor
   has_many :testruns, as: :user
   has_many :interventions, through: :user_exercise_interventions
   has_many :remote_evaluation_mappings, as: :user
   has_many :request_for_comments, as: :user
-  has_many :runners, as: :contributor
   has_many :events
   has_many :events_synchronized_editor, class_name: 'Event::SynchronizedEditor'
   has_many :pair_programming_exercise_feedbacks
   has_many :pair_programming_waiting_users
   has_one :codeharbor_link, dependent: :destroy
   accepts_nested_attributes_for :user_proxy_exercise_exercises
-
-  scope :with_submissions, -> { where('id IN (SELECT user_id FROM submissions)') }
 
   scope :in_study_group_of, lambda {|user|
                               unless user.admin?
@@ -45,18 +39,6 @@ class User < ApplicationRecord
                             }
 
   validates :platform_admin, inclusion: [true, false]
-
-  def internal_user?
-    is_a?(InternalUser)
-  end
-
-  def external_user?
-    is_a?(ExternalUser)
-  end
-
-  def programming_group?
-    false
-  end
 
   def learner?
     return true if current_study_group_id.nil?
@@ -84,19 +66,6 @@ class User < ApplicationRecord
   def current_study_group_membership
     # We use `where(...).limit(1)` instead of `find_by(...)` to allow query chaining
     study_group_memberships.where(study_group: current_study_group_id).limit(1)
-  end
-
-  def to_s
-    displayname
-  end
-
-  def to_page_context
-    {
-      id:,
-      type: self.class.name,
-      consumer: consumer.name,
-      displayname:,
-    }
   end
 
   def self.find_by_id_with_type(id_with_type)
